refactor: extract conversation state into useChat hook
Moves message history, streaming state, submit, and cancel
logic out of App into a dedicated useChat hook. App becomes
a thin UI shell that delegates state management to the hook.

Closes #42
